Fixed bug with AdaptiveGridView where items were not being sized when initially grouped.
Problem: MeasureOverride is not called when switching to Grouped view, therefore MinItemWidth and MinItemHeight was ignored.
Fix: Use ArrangeOverride instead of MeasureOverride.

Other committed items:
* Added a Grouping demo to examples app
* Added a new attached property: IsHidden
* Because example app has grown in complexity, I've moved model classes to dedicated folder
* General code cleanup
