+++ UwpHelpers/UwpHelpers.Examples/Models/DemoPage.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace UwpHelpers.Examples.Models
+{
+    public class DemoPage
+    {
+        public Type DemoPageType { get; set; }
+        public string DemoTitle { get; set; }
+        public string GlyphIcon { get; set; }
+    }
+}
