+++ UwpHelpers/UwpHelpers.Examples/Models/IncrementableItem.cs
@@ -1,0 +1,8 @@
+﻿namespace UwpHelpers.Examples.Models
+{
+    public class IncrementableItem
+    {
+        public int Id { get; set; }
+        public string Title { get; set; }
+    }
+}
